Copy dependencies into a single directory (#22)
Copies dependencies into a single directory instead of one directory
per entry point. This avoids 'protoc' erroring out when a .proto
exists in multiple entry points. When the two files are not
identical, a warning is emitted.

Changes the semantics of the entry point name: Previously, it was
arbitrary. Now, the location to which the .protos are copied can
be (optionally) specified after a semicolon.

Always include the base google .proto files.

Co-authored-by: Alex Kaszynski <akascap@gmail.com>